backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
google-generativeai==0.8.5
supabase==2.18.1
python-dotenv==1.1.1
pillow==11.3.0